@@ -54,19 +54,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/platform-server": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@nguniversal/express-engine": "^8.2.6",
-    "@nguniversal/module-map-ngfactory-loader": "^8.2.6",
+    "@angular/animations": "^9.1.0",
+    "@angular/cdk": "^9.2.0",
+    "@angular/common": "^9.1.0",
+    "@angular/compiler": "^9.1.0",
+    "@angular/core": "^9.1.0",
+    "@angular/forms": "^9.1.0",
+    "@angular/material": "^9.2.0",
+    "@angular/platform-browser": "^9.1.0",
+    "@angular/platform-browser-dynamic": "^9.1.0",
+    "@angular/platform-server": "^9.1.0",
+    "@angular/router": "^9.1.0",
+    "@nguniversal/express-engine": "^9.1.0",
     "@types/express": "^4.17.2",
     "angulartics2": "^8.3.0",
     "better-npm-run": "^0.1.1",
@@ -88,6 +87,7 @@
     "rxjs": "^6.5.4",
     "sanitize-html": "^1.21.1",
     "spdy": "^4.0.1",
+    "tslib": "^1.10.0",
     "workbox-core": "^5.0.0",
     "workbox-routing": "^5.0.0",
     "workbox-strategies": "^5.0.0",
@@ -104,28 +104,26 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.25",
-    "@angular-devkit/build-ng-packagr": "^0.803.25",
-    "@angular/cli": "^8.3.25",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular-devkit/build-ng-packagr": "~0.901.0",
+    "@angular/cli": "^9.1.0",
+    "@angular/compiler-cli": "^9.1.0",
+    "@angular/language-service": "^9.1.0",
     "@scullyio/scully": "0.0.80",
     "@types/html-minifier": "^3.5.3",
     "@types/jasmine": "^3.4.6",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^12.7.12",
+    "@types/node": "^12.11.1",
     "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "local-web-server": "^3.0.7",
     "ncp": "^2.0.0",
-    "ng-packagr": "^5.7.1",
+    "ng-packagr": "^9.0.0",
     "ts-loader": "^6.2.1",
     "ts-node": "~8.3.0",
-    "tsickle": "0.37.0",
-    "tslib": "^1.10.0",
     "tslint": "^5.20.1",
-    "typescript": "~3.5.3",
+    "typescript": "~3.8.3",
     "webpack-bundle-analyzer": "^3.6.0",
     "webpack-cli": "^3.3.10",
     "webpack-node-externals": "^1.7.2",
