@@ -10,7 +10,7 @@
  */
 
 import { isPlatformBrowser } from '@angular/common';
-import { ClassProvider, FactoryProvider, InjectionToken, PLATFORM_ID } from '@angular/core';
+import { ClassProvider, FactoryProvider, InjectionToken, PLATFORM_ID, Injectable } from '@angular/core';
 
 /* Create a new injection token for injecting the window. */
 export const WINDOW = new InjectionToken('WindowToken');
@@ -23,6 +23,7 @@ export abstract class WindowRef {
 }
 
 /* Define class that implements the abstract class and returns the native window object. */
+@Injectable()
 export class BrowserWindowRef extends WindowRef {
   constructor() {
     super();
